Lots of changes. Parse + compile smart_input
Refactor grammar a little, Add AssignNode that consolidates all
assignments.

Unfortunately, there's a reduce/reduce conflict right now that needs
to be mitigated by the lexer, but the proper path is chosen.

Fix many small bugs relating to newlines and spaces.

Add new tests that revealed weaknesses in the parser or compiler.