change scope implementation, fix bugs related to scope
Scope is now its own object, and is able to merge itself with other
scopes and keep references to any functions that the scope belongs
to.

Fix bugs and add tests for cases where `for var in expr` blocks
weren't scoping `var` properly if nested within a function that also
has a `var` argument. Now, the most immediate scope takes precedent
and `var` stays `var` and does NOT become `a:var`. To access `a:var`
in this scenario, use `a:var` explicitly, just like in VimL.